mm/vmstat: Protect per cpu variables with preempt disable on RT
Disable preemption on -RT for the vmstat code. On vanila the code runs in
IRQ-off regions while on -RT it is not. "preempt_disable" ensures that the
same ressources is not updated in parallel due to preemption.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

(cherry picked from commit 3f5ecd6ebb43045ea420a7fad5d0148ba194d56f)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>