mm/vmstat: Protect per cpu variables with preempt disable on RT
Disable preemption on -RT for the vmstat code. On vanila the code runs in
IRQ-off regions while on -RT it is not. "preempt_disable" ensures that the
same ressources is not updated in parallel due to preemption.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

(cherry picked from commit 1b97f02a115a13e6cdec9e9400a3afdfd63a212a)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>