mm/vmstat: Protect per cpu variables with preempt disable on RT
Disable preemption on -RT for the vmstat code. On vanila the code runs in
IRQ-off regions while on -RT it is not. "preempt_disable" ensures that the
same ressources is not updated in parallel due to preemption.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

(cherry picked from commit 10ead13c88fe0f3733e5cfee71c0c0dce85611bc)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>