sched: Add saved_state for tasks blocked on sleeping locks
Spinlocks are state preserving in !RT. RT changes the state when a
task gets blocked on a lock. So we need to remember the state before
the lock contention. If a regular wakeup (not a RTmutex related
wakeup) happens, the saved_state is updated to running. When the lock
sleep is done, the saved state is restored.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 3f045bdd0695018e44ff92d7792c1f95eba3e180)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>