x86: kvm Require const tsc for RT
Non constant TSC is a nightmare on bare metal already, but with
virtualization it becomes a complete disaster because the workarounds
are horrible latency wise. That's also a preliminary for running RT in
a guest on top of a RT host.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 2ee03aa971e33a45373b11eda58238ea14eaa04e)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>