configs: New config in arch/s390 for v5.4-rc1
Message-id: <20190923135800.25646-1-jcline@redhat.com>
Patchwork-id: 272760
O-Subject: [ARK INTERNAL PATCHv2] [redhat] configs: New config in arch/s390 for v5.4-rc1
Bugzilla:
RH-Acked-by: Thomas Huth <thuth@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* arch/s390/Kconfig (help): MARCH_Z15 (bool)
    Select this to enable optimizations for IBM z15 (8562
    and 8561 series). The kernel will be slightly faster but will not
    work on older machines.
* arch/s390/Kconfig (help): TUNE_Z15 (bool)
    (no helpstring found)

Cc: Neil Horman <nhorman@redhat.com>
Cc: Thomas Huth <thuth@redhat.com>
Cc: Cornelia Huck <cohuck@redhat.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Jeremy Cline <jcline@redhat.com>