configs: New config in drivers/cpuidle for v5.4-rc1
Message-id: <20190923140215.26145-1-jcline@redhat.com>
Patchwork-id: 272761
O-Subject: [ARK INTERNAL PATCHv2] [redhat] configs: New config in drivers/cpuidle for v5.4-rc1
Bugzilla:
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/cpuidle/Kconfig.arm (help): ARM_PSCI_CPUIDLE (bool)
    Select this to enable PSCI firmware based CPUidle driver for ARM.
    It provides an idle driver that is capable of detecting and
    managing idle states through the PSCI firmware interface.
* drivers/cpuidle/Kconfig (help): CPU_IDLE_GOV_HALTPOLL (bool)
    This governor implements haltpoll idle state selection, to be
    used in conjunction with the haltpoll cpuidle driver, allowing
    for polling for a certain amount of time before entering idle
    state.

    Some virtualized workloads benefit from using it.
* drivers/cpuidle/Kconfig (help): HALTPOLL_CPUIDLE (tristate)
    This option enables halt poll cpuidle driver, which allows to poll
    before halting in the guest (more efficient than polling in the
    host via halt_poll_ns for some scenarios).

Cc: John Linville <linville@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>