configs: New config in arch/arm64 for v5.4-rc1
Message-id: <20190923141122.1264-1-jcline@redhat.com>
Patchwork-id: 272788
O-Subject: [ARK INTERNAL PATCHv2] [redhat] configs: New config in arch/arm64 for v5.4-rc1
Bugzilla:
RH-Acked-by: Jeremy Linton <jlinton@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* arch/arm64/Kconfig (help): ARM64_TAGGED_ADDR_ABI (bool)
    When this option is enabled, user applications can opt in to a
    relaxed ABI via prctl() allowing tagged addresses to be passed
    to system calls as pointer arguments. For details, see
    Documentation/arm64/tagged-address-abi.txt.
* arch/arm64/Kconfig (help): ARM64_VA_BITS_52 (bool)
    Enable 52-bit virtual addressing for userspace when explicitly
    requested via a hint to mmap(). The kernel will also use 52-bit
    virtual addresses for its own mappings (provided HW support for
    this feature is available, otherwise it reverts to 48-bit).

    NOTE: Enabling 52-bit virtual addressing in conjunction with
    ARMv8.3 Pointer Authentication will result in the PAC being
    reduced from 7 bits to 3 bits, which may have a significant
    impact on its susceptibility to brute-force attacks.

    If unsure, select 48-bit virtual addressing instead.

Cc: Mark Salter <msalter@redhat.com>
Cc: Jeremy Linton <jlinton@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Cc: Andrew Jones <drjones@redhat.com>
Cc: Wei Huang <wei@redhat.com>
Cc: Eric Auger <eric.auger@redhat.com>
Cc: rhvirt-patches@redhat.com
Signed-off-by: Jeremy Cline <jcline@redhat.com>