configs: New config in drivers/edac for v5.4-rc1
Message-id: <20191001153139.14526-1-jcline@redhat.com>
Patchwork-id: 275492
O-Subject: [ARK INTERNAL PATCHv2] [redhat] configs: New config in drivers/edac for v5.4-rc1
Bugzilla:
RH-Acked-by: Aristeu Rozanski <aris@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/edac/Kconfig (help): EDAC_BLUEFIELD (tristate)
    Support for error detection and correction on the
    Mellanox BlueField SoCs.

Cc: Aristeu Rozanski <aris@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>