configs: New config in drivers/pinctrl for v5.4-rc1
Message-id: <20191014162300.16308-15-jcline@redhat.com>
Patchwork-id: 278885
O-Subject: [ARK INTERNAL PATCH] configs: New config in drivers/pinctrl for v5.4-rc1
Bugzilla:
RH-Acked-by: Laura Abbott <labbott@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* drivers/pinctrl/qcom/Kconfig (help): PINCTRL_SC7180 (tristate)
    This is the pinctrl, pinmux, pinconf and gpiolib driver for the
    Qualcomm Technologies Inc TLMM block found on the Qualcomm
    Technologies Inc SC7180 platform.

Signed-off-by: Jeremy Cline <jcline@redhat.com>