configs: New config in fs/fuse for v5.4-rc1
Message-id: <20191104195711.611286-1-jcline@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

Upstream Status: RHEL only.

Hi,

As part of the rebase effort to v5.4-rc1, New config option(s) were
exposed. Please review the options below.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

* fs/fuse/Kconfig (help): VIRTIO_FS (tristate)
    The Virtio Filesystem allows guests to mount file systems from the
    host.

    If you want to share files between guests or with the host, answer Y
    or M.

Cc: Brian Foster <bfoster@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>