[redhat] Be smarter about getting rid of debuginfo
We've been doing a workaround of using a uuid for the build-id for
userspace binaries to let us get parallel debuginfo (the trick we
use for the kernel doesn't extend to userspace binaries). This
works but isn't ideal for reproducibility. Given we don't actually
care about this debuginfo, just remove it from the buildroot completely.

This is a back port of a patch in Fedora's kernel specfile.

Cc: Don Zickus <dzickus@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Jan Stancek <jstancek@redhat.com>
Cc: Augusto Caringi <acaringi@redhat.com>
Cc: Frantisek Hrbata <fhrbata@redhat.com>
Cc: Luis Claudio Goncalves <lgoncalv@redhat.com>
Cc: Juri Lelli <jlelli@redhat.com>
Cc: Rado Vrbovsky <rvrbovsk@redhat.com>
Cc: Herton R. Krzesinski <herton@redhat.com>
Cc: Denys Vlasenko <dvlasenk@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Signed-off-by: Jeremy Cline <jcline@redhat.com>