Add a script to generate configuration patches
After generating an initial set of commits in the pending directory,
we need to split that out into individual commits for review. Add
a script to do so. This script is designed to be run from the
root of the kernel tree so we can run get_maintainers.pl if the
appropriate maintainers file is available.

Signed-off-by: Laura Abbott <labbott@redhat.com>