[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_COMPAT_32BIT_TIME:

 This enables 32 bit time_t support in addition to 64 bit time_t support.
 This is relevant on all 32-bit architectures, and 64-bit architectures
 as part of compat syscall handling.

 Symbol: COMPAT_32BIT_TIME [=y]
 Type  : bool
 Prompt: Provide system calls for 32-bit time_t
   Location:
     -> General architecture-dependent options
   Defined at arch/Kconfig:799

---