[redhat] New configs in arch/s390
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_S390_UNWIND_SELFTEST:

 This option enables s390 specific stack unwinder testing kernel
 module. This option is not useful for distributions or general
 kernels, but only for kernel developers working on architecture code.

 Say N if you are unsure.

 Symbol: S390_UNWIND_SELFTEST [=n]
 Type  : tristate
 Prompt: Test unwind functions
   Location:
     -> Selftests
   Defined at arch/s390/Kconfig:1022

---

Cc: Thomas Huth <thuth@redhat.com>