[redhat] New configs in fs/nfsd
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_NFSD_V4_2_INTER_SSC:

 This option enables support for NFSv4.2 inter server to
 server copy where the destination server calls the NFSv4.2
 client to read the data to copy from the source server.

 If unsure, say N.

 Symbol: NFSD_V4_2_INTER_SSC [=n]
 Type  : bool
 Defined at fs/nfsd/Kconfig:137
   Prompt: NFSv4.2 inter server to server COPY
   Depends on: NETWORK_FILESYSTEMS [=y] && NFSD_V4 [=y] && NFS_V4_1 [=y] && NFS_V4_2 [=y]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])

---

Cc: Red Hat NFS Team <nfs-team@redhat.com>