[redhat] New configs in drivers/video
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BACKLIGHT_LED:

 If you have a LCD backlight adjustable by LED class driver, say Y
 to enable this driver.

 Symbol: BACKLIGHT_LED [=n]
 Type  : tristate
 Defined at drivers/video/backlight/Kconfig:459
   Prompt: Generic LED based Backlight Driver
   Depends on: HAS_IOMEM [=y] && BACKLIGHT_CLASS_DEVICE [=y] && LEDS_CLASS [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Graphics support
         -> Backlight & LCD device support
           -> Lowlevel Backlight controls (BACKLIGHT_CLASS_DEVICE [=y])

Lyude isn't aware of anything that uses this currently, but noted that
that it's likely it will be needed for newer laptops and so recommended
"m".

---

Cc: Lyude Paul <lyude@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>