[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PCIE_EDR:

 This option adds Error Disconnect Recover support as specified
 in the Downstream Port Containment Related Enhancements ECN to
 the PCI Firmware Specification r3.2.  Enable this if you want to
 support hybrid DPC model which uses both firmware and OS to
 implement DPC.

 Symbol: PCIE_EDR [=n]
 Type  : bool
 Defined at drivers/pci/pcie/Kconfig:145
   Prompt: PCI Express Error Disconnect Recover support
   Depends on: PCI [=y] && PCIE_DPC [=y] && ACPI [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])

Enabled at the request of Myron Stowe.

---

 CONFIG_PCIE_LAYERSCAPE_GEN4:

 Say Y here if you want PCIe Gen4 controller support on
 Layerscape SoCs.
 Symbol: PCIE_LAYERSCAPE_GEN4 [=n]
 Type  : bool
 Defined at drivers/pci/controller/mobiveil/Kconfig:25
   Prompt: Freescale Layerscape PCIe Gen4 controller
   Depends on: PCI [=y] && OF [=y] && (ARM64 [=y] || ARCH_LAYERSCAPE [=n]) && PCI_MSI_IRQ_DOMAIN [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI controller drivers
           -> Mobiveil PCIe Core Support
 Selects: PCIE_MOBIVEIL_HOST [=n]

---

Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>