[redhat] New configs in drivers/pinctrl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PINCTRL_IPQ6018:

 This is the pinctrl, pinmux, pinconf and gpiolib driver for
 the Qualcomm Technologies Inc. TLMM block found on the
 Qualcomm Technologies Inc. IPQ6018 platform. Select this for
 IPQ6018.

 Symbol: PINCTRL_IPQ6018 [=n]
 Type  : tristate
 Defined at drivers/pinctrl/qcom/Kconfig:53
   Prompt: Qualcomm Technologies, Inc. IPQ6018 pin controller driver
   Depends on: PINCTRL [=y] && (ARCH_QCOM [=y] || COMPILE_TEST [=n]) && GPIOLIB [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Pin controllers (PINCTRL [=y])
 Selects: PINCTRL_MSM [=y]

---