[redhat] New configs in drivers/vhost
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_VHOST_VDPA:

 This kernel module can be loaded in host kernel to accelerate
 guest virtio devices with the vDPA-based backends.

 To compile this driver as a module, choose M here: the module
 will be called vhost_vdpa.

 Symbol: VHOST_VDPA [=n]
 Type  : tristate
 Defined at drivers/vhost/Kconfig:62
   Prompt: Vhost driver for vDPA-based backend
   Depends on: VHOST_MENU [=y] && EVENTFD [=y]
   Location:
     -> Device Drivers
       -> VHOST drivers (VHOST_MENU [=y])
 Selects: VHOST [=m] && VDPA [=n]

Enabled at the request of Michael S. Tsirkin. Disabled on zfcpdump as it
seems unnecessary.

---

Cc: "Michael S. Tsirkin" <mst@redhat.com>
Cc: Jason Wang <jasowang@redhat.com>
Cc: rhvirt-patches@redhat.com