Exit non-zero if the tag already exists for a release
The CI job can't determine if the release already exists or not with
the current script, so it fails when it tries to run the follow-up
scripts. This allows us to skip the rest of the job when the tag already
exists so maintainers only see failed jobs when something is actually
wrong.

If the script exits 1 or 2 it's for reasons that require maintainer
intervention. This exits 3 to distinguish between those cases and the
case that there hasn't been an update to upstream.

Signed-off-by: Jeremy Cline <jcline@redhat.com>