[redhat] New configs in drivers/watchdog
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_SMC_WATCHDOG:

 Say Y here to include support for a watchdog timer
 implemented by the EL3 Secure Monitor on ARM platforms.
 Requires firmware support.
 To compile this driver as a module, choose M here: the
 module will be called arm_smc_wdt.

 Symbol: ARM_SMC_WATCHDOG [=n]
 Type  : tristate
 Defined at drivers/watchdog/Kconfig:871
   Prompt: ARM Secure Monitor Call based watchdog support
   Depends on: WATCHDOG [=y] && (ARM || ARM64 [=y]) && OF [=y] && HAVE_ARM_SMCCC [=y]
   Location:
     -> Device Drivers
       -> Watchdog Timer Support (WATCHDOG [=y])
 Selects: WATCHDOG_CORE [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>