New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_HIST_TRIGGERS_DEBUG:

 Add "hist_debug" file for each event, which when read will
 dump out a bunch of internal details about the hist triggers
 defined on that event.

 The hist_debug file serves a couple of purposes:

   - Helps developers verify that nothing is broken.

   - Provides educational information to support the details
     of the hist trigger internals as described by
     Documentation/trace/histogram-design.rst.

 The hist_debug output only covers the data structures
 related to the histogram definitions themselves and doesn't
 display the internals of map buckets or variable values of
 running histograms.

 If unsure, say N.

 Symbol: HIST_TRIGGERS_DEBUG [=n]
 Type  : bool
 Defined at kernel/trace/Kconfig:850
   Prompt: Hist trigger debug support
   Depends on: TRACING_SUPPORT [=y] && FTRACE [=y] && HIST_TRIGGERS [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])

Cc: Jerome Marchand <jmarchan@redhat.com>