drm/i915/gt: Only disable interrupts for the timeline lock on !force-threaded
According to commit
    d67739268cf0e ("drm/i915/gt: Mark up the nested engine-pm timeline lock as irqsafe")

the intrrupts are disabled the code may be called from an interrupt
handler and from preemptible context.
With `force_irqthreads' set the timeline mutex is never observed in IRQ
context so it is not neede to disable interrupts.

Disable only interrupts if not in `force_irqthreads' mode.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
(cherry picked from commit 24cff14d3791643ebe81cf7e23cd51c6402a780d)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>