New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_KERNEL_ZSTD:

 ZSTD is a compression algorithm targeting intermediate compression
 with fast decompression speed. It will compress better than GZIP and
 decompress around the same speed as LZO, but slower than LZ4. You
 will need at least 192 KB RAM or more for booting. The zstd command
 line tool is required for compression.

 Symbol: KERNEL_ZSTD [=n]
 Type  : bool
 Defined at init/Kconfig:282
   Prompt: ZSTD
   Depends on: <choice> && HAVE_KERNEL_ZSTD [=y]
   Location:
     -> General setup
       -> Kernel compression mode (<choice> [=y])

Cc: Prarit Bhargava <prarit@redhat.com>