u64_stats: Disable preemption on 32bit-UP/SMP with RT during updates
On RT the seqcount_t is required even on UP because the softirq can be
preempted. The IRQ handler is threaded so it is also preemptible.

Disable preemption on 32bit-RT during value updates. There is no need to
disable interrupts on RT because the handler is run threaded. Therefore
disabling preemption is enough to guarantee that the update is not
interruped.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
(cherry picked from commit a191d2f617c965bae7e71df7e70f67c5bf7a3592)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>