locking/rtmutex: Remove cruft
Most of this is around since the very beginning. I'm not sure if this
was used while the rtmutex-deadlock-tester was around but today it seems
to only waste memory:
- save_state: No users
- name: Assigned and printed if a dead lock was detected. I'm keeping it
  but want to point out that lockdep has the same information.
- file + line: Printed if ::name was NULL. This is only used for
  in-kernel locks so it ::name shouldn't be NULL and then ::file and
  ::line isn't used.
- magic: Assigned to NULL by rt_mutex_destroy().

Remove members of rt_mutex which are not used.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
(cherry picked from commit 79f095a728deab90d44f7efa32c21bd95061ab91)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>