Combine Red Hat patches into single patch
This in spirit reverts 0409b218390b564c44dd0181c5d0fe177d4c6bc3
and converts the broken out Red Hat patches back into a single diff.

The original idea was to make it easy for the Fedora community to see
what changes Red Hat was making on top of upstream's tarball.  The
concept was good, the reality with a source git tree had a flaw.

While 'git merging' with upstream, various conflicts were resolved
through git or the merge.  However, when broken out into individual
patches and placed on top of a tarball, this becomes a rebase.

The conflicts could not be resolved with later patches as the initial
patch was a problem.  The only solution was to periodically rebase
the source git tree which dampens the spirit of the idea of source-git.

This patch takes another apporach to solve the original problem:
- create a diff from upstream as a single patch
- create a Patchlist.changelog file to list the individual patches
  and their shas1 as seen in the source git tree.

This is what was there original and hence the partial revert (though
that patch went through many changes since then).

It isn't perfect, but we are still evolving.  This should resolve the
CKI build failures they continuously run into.

Signed-off-by: Don Zickus <dzickus@redhat.com>