Fix path location for ark-update-configs.sh
The ark-update-configs.sh script is written to be executed from the top of the
git tree.  However, from a make -C redhat/ command it is executed from the
redhat/ path.  This breaks the script when it needs to generate new configs
(redhat/gen_config_patches.sh).

A simple fix is to execute the script inside the redhat/Makefile from the top
level.  The alternative is to teach the script to be location agnostic, but
there is a handful of hardcoded places across 2 scripts that I thought it
wasn't worth it.

Tested by running the 'scheduled' merge-upstream job using my test-os-build
branch where it passed
(https://gitlab.com/cki-project/kernel-ark/-/pipelines/199415525)

Signed-off-by: Don Zickus <dzickus@redhat.com>