[redhat] New configs in drivers/net/ethernet/chelsio
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CHELSIO_INLINE_CRYPTO:

 Enable support for inline crypto.
 Allows enable/disable from list of inline crypto drivers.

 Symbol: CHELSIO_INLINE_CRYPTO [=y]
 Type  : bool
 Defined at drivers/net/ethernet/chelsio/inline_crypto/Kconfig:6
   Prompt: Chelsio Inline Crypto support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_CHELSIO [=y] && CHELSIO_T4 [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Chelsio devices (NET_VENDOR_CHELSIO [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>