[redhat] New configs in lib/Kconfig.kasan
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_TEST_KASAN_MODULE:

 This is a part of the KASAN test suite that is incompatible with
 KUnit. Currently includes tests that do bad copy_from/to_user
 accesses.

 Symbol: TEST_KASAN_MODULE [=n]
 Type  : tristate
 Defined at lib/Kconfig.kasan:181
   Prompt: KUnit-incompatible tests of KASAN bug detection capabilities
   Depends on: m && MODULES [=y] && KASAN [=y]
   Location:
     -> Kernel hacking
       -> Memory Debugging
         -> KASAN: runtime memory debugger (KASAN [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>