[redhat] New configs in drivers/soc
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_QCOM_CPR:

 Say Y here to enable support for the CPR hardware found on Qualcomm
 SoCs like QCS404.

 This driver populates CPU OPPs tables and makes adjustments to the
 tables based on feedback from the CPR hardware. If you want to do
 CPUfrequency scaling say Y here.

 To compile this driver as a module, choose M here: the module will
 be called qcom-cpr

 Symbol: QCOM_CPR [=n]
 Type  : tristate
 Defined at drivers/soc/qcom/Kconfig:29
   Prompt: QCOM Core Power Reduction (CPR) support
   Depends on: ARCH_QCOM [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> SOC (System On Chip) specific Drivers
         -> Qualcomm SoC drivers
 Selects: PM_OPP [=n] && REGMAP [=y]

---

Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>