highmem: Don't disable preemption on RT in kmap_atomic()
kmap_atomic() disables preemption and pagefaults for historical
reasons. The conversion to kmap_local(), which only disables
migration, cannot be done wholesale because quite some call sites need
to be updated to accommodate with the changed semantics.

On PREEMPT_RT enabled kernels the kmap_atomic() semantics are
problematic due to the implicit disabling of preemption which makes it
impossible to acquire 'sleeping' spinlocks within the kmap atomic
sections.

PREEMPT_RT replaces the preempt_disable() with a migrate_disable() for
more than a decade. It could be argued that this is a justification to
do this unconditionally, but PREEMPT_RT covers only a limited number of
architectures and it disables some functionality which limits the
coverage further.

Limit the replacement to PREEMPT_RT for now.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
https://lkml.kernel.org/r/20210810091116.pocdmaatdcogvdso@linutronix.de
(cherry picked from commit 477445fe0e4423fb07302a762f0230532d259970)
Signed-off-by: Juri Lelli <juri.lelli@redhat.com>