[redhat] New configs in fs/nfs
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_NFS_V4_2_READ_PLUS:

 This is intended for developers only. The READ_PLUS operation has
 been shown to have issues under specific conditions and should not
 be used in production.

 Symbol: NFS_V4_2_READ_PLUS [=n]
 Type  : bool
 Defined at fs/nfs/Kconfig:209
   Prompt: NFS: Enable support for the NFSv4.2 READ_PLUS operation
   Depends on: NETWORK_FILESYSTEMS [=y] && NFS_V4_2 [=y]
   Location:
     -> File systems
       -> Network File Systems (NETWORK_FILESYSTEMS [=y])

---

Cc: kernel-patches@redhat.com
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>