[redhat] New configs in drivers/thunderbolt
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_USB4_DMA_TEST:

 This allows sending and receiving DMA traffic through loopback
 connection. Loopback connection can be done by either special
 dongle that has TX/RX lines crossed, or by simply connecting a
 cable back to the host. Only enable this if you know what you
 are doing. Normal users and distro kernels should say N here.

 To compile this driver a module, choose M here. The module will be
 called thunderbolt_dma_test.

 Symbol: USB4_DMA_TEST [=n]
 Type  : tristate
 Defined at drivers/thunderbolt/Kconfig:34
   Prompt: DMA traffic test driver
   Depends on: USB4 [=y] && DEBUG_FS [=y]
   Location:
     -> Device Drivers
       -> Unified support for USB4 and Thunderbolt (USB4 [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>