[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_QCOM_GPI_DMA:

 Enable support for the QCOM GPI DMA controller. This controller
 provides DMA capabilities for a variety of peripheral buses such
 as I2C, UART, and SPI. By using GPI dmaengine driver, bus drivers
 can use a standardize interface that is protocol independent to
 transfer data between DDR and peripheral.

 Symbol: QCOM_GPI_DMA [=n]
 Type  : tristate
 Defined at drivers/dma/qcom/Kconfig:22
   Prompt: Qualcomm Technologies GPI DMA support
   Depends on: DMADEVICES [=y] && ARCH_QCOM [=y]
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=n]

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>