[redhat] New configs in kernel/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_DMA_MAP_BENCHMARK:

 Provides /sys/kernel/debug/dma_map_benchmark that helps with testing
 performance of dma_(un)map_page.

 See tools/testing/selftests/dma/dma_map_benchmark.c

 Symbol: DMA_MAP_BENCHMARK [=n]
 Type  : bool
 Defined at kernel/dma/Kconfig:228
   Prompt: Enable benchmarking of streaming DMA mapping
   Depends on: DEBUG_FS [=y]
   Location:
     -> Library routines

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>