[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SPI_AMD:

 Enables SPI controller driver for AMD SoC.

 Symbol: SPI_AMD [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:917
   Prompt: AMD SPI controller
   Depends on: SPI [=y] && SPI_MASTER [=y] && (SPI_MASTER [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])

---

Signed-off-by: CKI@GitLab <cki-project@redhat.com>