[redhat] spec: package decompressor vmlinux for s390
On s390 more and more code from the early boot stage is moved to the
decompressor. With this the code complexity and thus the chances to
introduce bugs increases. In order to be able to debug these early boot
bugs package the decompressors vmlinux together with the other debuginfo
for s390.

This change requires kernel commit 7a84ffc0471f ("s390/boot: add
build-id to decompressor") to work.

Upstream Status: RHEL only

Signed-off-by: Philipp Rudo <prudo@redhat.com>