[redhat] New configs in drivers/leds
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_LEDS_TRIGGER_TTY:

 This allows LEDs to be controlled by activity on ttys which includes
 serial devices like /dev/ttyS0.

 When build as a module this driver will be called ledtrig-tty.

 Symbol: LEDS_TRIGGER_TTY [=n]
 Type  : tristate
 Defined at drivers/leds/trigger/Kconfig:147
   Prompt: LED Trigger for TTY devices
   Depends on: NEW_LEDS [=y] && LEDS_TRIGGERS [=y] && TTY [=y]
   Location:
     -> Device Drivers
       -> LED Support (NEW_LEDS [=y])
         -> LED Trigger support (LEDS_TRIGGERS [=y])

---

Cc: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>