[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_MT7921E:

 This adds support for MT7921E 802.11ax 2x2:2SS wireless devices.

 To compile this driver as a module, choose M here.

 Symbol: MT7921E [=n]
 Type  : tristate
 Defined at drivers/net/wireless/mediatek/mt76/mt7921/Kconfig:2
   Prompt: MediaTek MT7921E (PCIe) support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_MEDIATEK [=y] && MAC80211 [=m] && PCI [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> MediaTek devices (WLAN_VENDOR_MEDIATEK [=y])
 Selects: MT76_CONNAC_LIB [=n] && WANT_DEV_COREDUMP [=y]

---

Cc: Jarod Wilson <jarod@redhat.com>
Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>