[redhat] New configs in drivers/hwmon
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SENSORS_AHT10:

 If you say yes here, you get support for the Aosong AHT10
 temperature and humidity sensors

 This driver can also be built as a module. If so, the module
 will be called aht10.

 Symbol: SENSORS_AHT10 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:247
   Prompt: Aosong AHT10
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])

---

 CONFIG_SENSORS_TPS23861:

 If you say yes here you get support for Texas Instruments
 TPS23861 802.3at PoE PSE chips.

 This driver can also be built as a module. If so, the module
 will be called tps23861.

 Symbol: SENSORS_TPS23861 [=n]
 Type  : tristate
 Defined at drivers/hwmon/Kconfig:1136
   Prompt: Texas Instruments TPS23861 PoE PSE
   Depends on: HWMON [=y] && I2C [=y]
   Location:
     -> Device Drivers
       -> Hardware Monitoring support (HWMON [=y])
 Selects: REGMAP_I2C [=m]

---

Cc: Dean Nelson <dnelson@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>