[redhat] New configs in drivers/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_INTEL_LDMA:

 Enable support for Intel Lightning Mountain SOC DMA controllers.
 These controllers provide DMA capabilities for a variety of on-chip
 devices such as HSNAND and GSWIP (Gigabit Switch IP).

 Symbol: INTEL_LDMA [=n]
 Type  : bool
 Defined at drivers/dma/lgm/Kconfig:2
   Prompt: Lightning Mountain centralized DMA controllers
   Depends on: DMADEVICES [=y] && (X86 [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> DMA Engine support (DMADEVICES [=y])
 Selects: DMA_ENGINE [=y] && DMA_VIRTUAL_CHANNELS [=y]

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>