[redhat] New configs in drivers/pci
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_PCIE_MICROCHIP_HOST:

 Say Y here if you want kernel to support the Microchip AXI PCIe
 Host Bridge driver.

 Symbol: PCIE_MICROCHIP_HOST [=n]
 Type  : bool
 Defined at drivers/pci/controller/Kconfig:278
   Prompt: Microchip AXI PCIe host bridge support
   Depends on: PCI [=y] && PCI_MSI [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> PCI controller drivers
 Selects: PCI_MSI_IRQ_DOMAIN [=y] && GENERIC_MSI_IRQ_DOMAIN [=y] && PCI_HOST_COMMON [=n]

---

Cc: Myron Stowe <myron.stowe@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>