[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_RANDOMIZE_KSTACK_OFFSET_DEFAULT:

 The kernel stack offset can be randomized (after pt_regs) by
 roughly 5 bits of entropy, frustrating memory corruption
 attacks that depend on stack address determinism or
 cross-syscall address exposures. This feature is controlled
 by kernel boot param "randomize_kstack_offset=on/off", and this
 config chooses the default boot state.

 Symbol: RANDOMIZE_KSTACK_OFFSET_DEFAULT [=n]
 Type  : bool
 Defined at arch/Kconfig:1114
   Prompt: Randomize kernel stack offset on syscall entry
   Depends on: HAVE_ARCH_RANDOMIZE_KSTACK_OFFSET [=y]
   Location:
     -> General architecture-dependent options

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>