[redhat] New configs in drivers/spi
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_SPI_ALTERA_CORE:

 "The core code for the Altera SPI Controller"

 Symbol: SPI_ALTERA_CORE [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:67
   Prompt: Altera SPI Controller core code
   Depends on: SPI [=y] && SPI_MASTER [=y]
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])
 Selects: REGMAP [=y]
 Selected by [n]:
   - SPI_ALTERA [=n] && SPI [=y] && SPI_MASTER [=y]
   - SPI_ALTERA_DFL [=n] && SPI [=y] && SPI_MASTER [=y] && FPGA_DFL [=n]

---

 CONFIG_SPI_HISI_KUNPENG:

 This enables support for HiSilicon SPI controller found on
 Kunpeng SoCs.

 This driver can also be built as a module. If so, the module
 will be called hisi-kunpeng-spi.

 Symbol: SPI_HISI_KUNPENG [=n]
 Type  : tristate
 Defined at drivers/spi/Kconfig:351
   Prompt: HiSilicon SPI Controller for Kunpeng SoCs
   Depends on: SPI [=y] && SPI_MASTER [=y] && (ARM64 [=y] && ACPI [=y] || COMPILE_TEST [=n])
   Location:
     -> Device Drivers
       -> SPI support (SPI [=y])

---

Cc: Al Stone <ahs3@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>