[redhat] New configs in drivers/bluetooth
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BT_VIRTIO:

 Virtio Bluetooth support driver.
 This driver supports Virtio Bluetooth devices.

 Say Y here to compile support for HCI over Virtio into the
 kernel or say M to compile as a module.

 Symbol: BT_VIRTIO [=n]
 Type  : tristate
 Defined at drivers/bluetooth/Kconfig:428
   Prompt: Virtio Bluetooth driver
   Depends on: NET [=y] && BT [=m] && VIRTIO [=y]
   Location:
     -> Networking support (NET [=y])
       -> Bluetooth subsystem support (BT [=m])
         -> Bluetooth device drivers

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>