[redhat] New configs in drivers/power
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_BATTERY_GOLDFISH:

 Say Y to enable support for the Goldfish battery and AC power
 driver. Originated in the Android Studio Emulator (goldfish) it is
 going to be used in other emulators.

 Symbol: BATTERY_GOLDFISH [=n]
 Type  : tristate
 Defined at drivers/power/supply/Kconfig:705
   Prompt: Goldfish battery driver
   Depends on: POWER_SUPPLY [=y] && HAS_IOMEM [=y]
   Location:
     -> Device Drivers
       -> Power supply class support (POWER_SUPPLY [=y])

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>