[redhat] New configs in drivers/cxl
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_CXL_MEM:

 The CXL.mem protocol allows a device to act as a provider of
 "System RAM" and/or "Persistent Memory" that is fully coherent
 as if the memory was attached to the typical CPU memory
 controller.

 Say 'y/m' to enable a driver (named "cxl_mem.ko" when built as
 a module) that will attach to CXL.mem devices for
 configuration, provisioning, and health monitoring. This
 driver is required for dynamic provisioning of CXL.mem
 attached memory which is a prerequisite for persistent memory
 support. Typically volatile memory is mapped by platform
 firmware and included in the platform memory map, but in some
 cases the OS is responsible for mapping that memory. See
 Chapter 2.3 Type 3 CXL Device in the CXL 2.0 specification.

 If unsure say 'm'.

 Symbol: CXL_MEM [=n]
 Type  : tristate
 Defined at drivers/cxl/Kconfig:16
   Prompt: CXL.mem: Memory Devices
   Depends on: CXL_BUS [=y]
   Location:
     -> Device Drivers
       -> PCI support (PCI [=y])
         -> CXL (Compute Express Link) Devices Support (CXL_BUS [=y])

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 - CONFIG_CXL_MEM set to module
v3:
 - CONFIG_CXL_MEM_RAW_COMMANDS disabled

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>