rpmspec: fix check that calls InitBuildVars
InitBuildVars is only called in case BuildKernel is not called (because
it already calls InitBuildVars), and that is done eg. when all kernel
builds are disabled but still building of tools is wanted which needs
the definitions/variables it declares. However, the check is missing the
case when armv7hl with_pae is enabled which already called
BuildKernel->InitBuildVars, this just adds the missing check.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>