[redhat] New configs in security/integrity
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_IMA_DISABLE_HTABLE:

 This option disables htable to allow measurement of duplicate records.

 Symbol: IMA_DISABLE_HTABLE [=n]
 Type  : bool
 Defined at security/integrity/ima/Kconfig:338
   Prompt: Disable htable to allow measurement of duplicate records
   Depends on: INTEGRITY [=y] && IMA [=y]
   Location:
     -> Security options
       -> Integrity subsystem (INTEGRITY [=y])

---

Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>