[redhat] New configs in drivers/firmware
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 CONFIG_ARM_FFA_TRANSPORT:

 This Firmware Framework(FF) for Arm A-profile processors describes
 interfaces that standardize communication between the various
 software images which includes communication between images in
 the Secure world and Normal world. It also leverages the
 virtualization extension to isolate software images provided
 by an ecosystem of vendors from each other.

 This driver provides interface for all the client drivers making
 use of the features offered by ARM FF-A.

 Symbol: ARM_FFA_TRANSPORT [=n]
 Type  : tristate
 Defined at drivers/firmware/arm_ffa/Kconfig:2
   Prompt: Arm Firmware Framework for Armv8-A
   Depends on: OF [=y] && ARM64 [=y]
   Location:
     -> Firmware Drivers

---

Cc: Lenny Szubowicz <lszubowi@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Langsdorf <mlangsdo@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>