eventfd: Make recursion protection RT safe
The eventfd recursion protection uses a per CPU counter under the
assumption that preemption is disabled. On RT this is not the case so if a
task preempts the recursion protected region and invokes eventfd_signal()
it triggers the recursion warning and the signaling is skipped.

Paolo suggested to protect the recursion count with a local lock, but the
recursion protection is really per task and not per CPU.

Replace the per CPU counter, which is in fact a boolean flag, with a
status bit in task struct.

Reported-by: Daniel Bristot de Oliveira <bristot@redhat.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>