[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_SM4_AESNI_AVX2_X86_64 [=n]
 Type  : tristate
 Defined at crypto/Kconfig:1594
   Prompt: SM4 cipher algorithm (x86_64/AES-NI/AVX2)
   Depends on: CRYPTO [=y] && X86 [=y] && 64BIT [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_SIMD [=y] && CRYPTO_ALGAPI [=y] && CRYPTO_LIB_SM4 [=n] && CRYPTO_SM4_AESNI_AVX_X86_64 [=n]

---

 Symbol: CRYPTO_SM4_AESNI_AVX_X86_64 [=n]
 Type  : tristate
 Defined at crypto/Kconfig:1573
   Prompt: SM4 cipher algorithm (x86_64/AES-NI/AVX)
   Depends on: CRYPTO [=y] && X86 [=y] && 64BIT [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_SKCIPHER [=y] && CRYPTO_SIMD [=y] && CRYPTO_ALGAPI [=y] && CRYPTO_LIB_SM4 [=n]
 Selected by [n]:
   - CRYPTO_SM4_AESNI_AVX2_X86_64 [=n] && CRYPTO [=y] && X86 [=y] && 64BIT [=y]

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Cc: Vladis Dronov <vdronov@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>