[redhat] New configs in drivers/dma-buf
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DMABUF_SYSFS_STATS [=n]
 Type  : bool
 Defined at drivers/dma-buf/Kconfig:75
   Prompt: DMA-BUF sysfs statistics
   Location:
     -> Device Drivers
       -> DMABUF options
 Selects: DMA_SHARED_BUFFER [=y]

---

Cc: Lyude Paul <lyude@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>