[redhat] New configs in security/Kconfig.hardening
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: ZERO_CALL_USED_REGS [=n]
 Type  : bool
 Defined at security/Kconfig.hardening:235
   Prompt: Enable register zeroing on function exit
   Depends on: CC_HAS_ZERO_CALL_USED_REGS [=y]
   Location:
     -> Security options
       -> Kernel hardening options
         -> Memory initialization

---

Cc: Ondrej Mosnacek <omosnace@redhat.com>
Cc: Christoph von Recklinghausen <crecklin@redhat.com>
Cc: Richard Guy Briggs <rbriggs@redhat.com>
Cc: Vladis Dronov <vdronov@redhat.com>
Cc: Wander Costa <wcosta@redhat.com>
Cc: Yannick Cote <ycote@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>