[redhat] New configs in kernel/dma
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DMA_RESTRICTED_POOL [=n]
 Type  : bool
 Defined at kernel/dma/Kconfig:83
   Prompt: DMA Restricted Pool
   Depends on: OF [=y] && OF_RESERVED_MEM [=y] && SWIOTLB [=y]
   Location:
     -> Library routines

---

Cc: Jerry Snitselaar <jsnitsel@redhat.com>
Cc: Donald Dutile <ddutile@redhat.com>
Cc: Clark Williams <williams@redhat.com>
Cc: Joel Savitz <jsavitz@redhat.com>
Cc: Lucas Zampieri <lzampier@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>