[redhat] New configs in arch/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: LTO_NONE [=y]
 Type  : bool
 Defined at arch/Kconfig:665
   Prompt: None
   Depends on: <choice>
   Location:
     -> General architecture-dependent options
       -> Link Time Optimization (LTO) (<choice> [=y])

---

Cc: Prarit Bhargava <prarit@redhat.com>
Cc: David Arcari <darcari@redhat.com>
Cc: Mark Salter <msalter@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>