[redhat] New configs in crypto/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: CRYPTO_CURVE25519 [=n]
 Type  : tristate
 Defined at crypto/Kconfig:285
   Prompt: Curve25519 algorithm
   Depends on: CRYPTO [=y]
   Location:
     -> Cryptographic API (CRYPTO [=y])
 Selects: CRYPTO_KPP [=y] && CRYPTO_LIB_CURVE25519_GENERIC [=m]
 Selected by [n]:
   - CRYPTO_DEV_HISI_HPRE [=n] && CRYPTO [=y] && CRYPTO_HW [=y] && PCI [=y] && PCI_MSI [=y] && (UACCE [=n] || UACCE [=n]=n) && (ARM64 || COMPILE_TEST [=n] && 64BIT [=y]) && ACPI [=y]

---

Cc: Herbert Xu <herbert.xu@redhat.com>
Cc: "David S. Miller" <davem@redhat.com>
Cc: Bruno Meneguele <bmeneg@redhat.com>
Cc: Ondrej Mosnacek <omosnace@redhat.com>
Cc: Phil Auld <pauld@redhat.com>
Cc: Vladis Dronov <vdronov@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>