Build CONFIG_SPI_PXA2XX as a module on x86
While turning off CONFIG_SPI_PXA2XX was the reviewed and set option for
RHEL, SND_SOC_INTEL_BDW_RT5677_MACH is built as a module in RHEL 8 and
was set that way for ark/RHEL 9. Unofrtunately that driver depends on
the SPI controller, and upstream commit 630db1549356f made it select
SPI_PXA2XX leaving us with a broken config. At some point, an entry was
made in pending-common to enable SPI_PXA2XX on x86, but there was no
corresponding MR to get review and make it official.  We have been
building this way for over a year, so let's review this and set it
properly.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>