spec: Keep .BTF section in modules
Currently .BTF section is moved to debuginfo objects,
adding _find_debuginfo_opts option to keep it.

The .BTF section needs to be in the module when it's loaded
so can't be in debug objects.

This increases size of each module. Currently there's known
dedup issue on s390x, that makes the size of all modules
double [1] so I'm not enabling it for s390x for the moment.

Also armv7hl arch does not have CONFIG_DEBUG_INFO_BTF enabled,
so this does not affect armv7hl.

Size of modules in total - kernel dir under /lib/modules/VER/
from kernel-core and kernel-module packages:

            base    new
  aarch64    64M    80M
  armv7hl    55M    55M
  ppc64le    56M    68M
  s390x      22M    22M
  x86_64     67M    82M

[1] https://lore.kernel.org/bpf/20211023120452.212885-1-jolsa@kernel.org/
Signed-off-by: Jiri Olsa <jolsa@redhat.com>