Enable CONFIG_QCOM_SCM for arm
Commit 0f0f80d9d5db0b9a78a6ec74714f08b7b8bda0de added
CONFIG_ARM_SMMU_QCOM which is a hidden config that defaults to on if
ARM_SMMU and ARCH_QCOM are enabled.   It also selects QCOM_SCM. As both
ARM_SMMU and ARCH_QCOM are enabled for RHEL configs, we need to flip
QCOM_SCM for arm as well of we get a mismatch.  Doing so will also
create a new config item QCOM_SCM_DOWNLOAD_MODE_DEFAULT, which I have
left at the default config of off. If that needs to change, I am happy
to change it.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>