[redhat] New configs in drivers/net/ethernet/asix
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NET_VENDOR_ASIX [=y]
 Type  : bool
 Defined at drivers/net/ethernet/asix/Kconfig:5
   Prompt: Asix devices
   Depends on: NETDEVICES [=y] && ETHERNET [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])

---

 Symbol: SPI_AX88796C [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/asix/Kconfig:14
   Prompt: Asix AX88796C-SPI support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_ASIX [=y] && SPI [=y] && GPIOLIB [=y]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Ethernet driver support (ETHERNET [=y])
           -> Asix devices (NET_VENDOR_ASIX [=y])
 Selects: PHYLIB [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>