[redhat] New configs in drivers/net/wireless
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: MT7921S [=n]
 Type  : tristate
 Defined at drivers/net/wireless/mediatek/mt76/mt7921/Kconfig:17
   Prompt: MediaTek MT7921S (SDIO) support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_MEDIATEK [=y] && MAC80211 [=m] && MMC [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> MediaTek devices (WLAN_VENDOR_MEDIATEK [=y])
 Selects: MT76_SDIO [=n] && MT7921_COMMON [=n]

---

 Symbol: RTW89 [=n]
 Type  : tristate
 Defined at drivers/net/wireless/realtek/rtw89/Kconfig:2
   Prompt: Realtek 802.11ax wireless chips support
   Depends on: NETDEVICES [=y] && WLAN [=y] && WLAN_VENDOR_REALTEK [=y] && MAC80211 [=m]
   Location:
     -> Device Drivers
       -> Network device support (NETDEVICES [=y])
         -> Wireless LAN (WLAN [=y])
           -> Realtek devices (WLAN_VENDOR_REALTEK [=y])

---

Cc: "Íñigo Huguet" <ihuguet@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: John Linville <linville@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>