kernel.spec: Add glibc-static build requirement
CKI and brew builds resulted in two different values for
CONFIG_CC_CAN_LINK_STATIC.  In CKI CONFIG_CC_CAN_LINK_STATIC=y while in
brew CONFIG_CC_CAN_LINK_STATIC was not defined.  This occurs because
glibc-static is installed on the CKI build containers and was not
installed in brew.

In order to guarantee consistency across build environments add a build
requirement for the glibc-static package.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>