[redhat] New configs in drivers/net/ethernet/engleder
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NET_VENDOR_ENGLEDER [=y]
 Type  : bool
 Defined at drivers/net/ethernet/engleder/Kconfig:6
   Prompt: Engleder devices
   Depends on: NETDEVICES [=y] && ETHERNET [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Ethernet driver support (ETHERNET [=y])

---

 Symbol: TSNEP [=n]
 Type  : tristate
 Defined at drivers/net/ethernet/engleder/Kconfig:19
   Prompt: TSN endpoint support
   Depends on: NETDEVICES [=y] && ETHERNET [=y] && NET_VENDOR_ENGLEDER [=y] && HAS_IOMEM [=y] && HAS_DMA [=y] && PTP_1588_CLOCK_OPTIONAL [=y]
   Location:
     Main menu
       -> Device Drivers
         -> Network device support (NETDEVICES [=y])
           -> Ethernet driver support (ETHERNET [=y])
             -> Engleder devices (NET_VENDOR_ENGLEDER [=y])
 Selects: PHYLIB [=y]

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>

v2:
 CONFIG_NET_VENDOR_ENGLEDER is disabled

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>