Add packaged but empty /lib/modules/<kver>/systemtap/
It was requested in rhbz 1477295 quite some time ago and got missed, but
a new bug has been opened against RHEL 9 requesting the same. This is the
default location for systemtap to pull in taps for stapusr executed
scripts.  While there are a number of ways to do this, the fact that the
directory is versioned with the kernel means it is likely best done in
the kernel itself.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>