spec: don't overwrite auto.conf with .config
This seems unnecessary (the auto.conf file is already there) and after
commit 129ab0d2d9f3 ("kbuild: do not quote string values in
include/config/auto.conf") it actually makes building external modules
fail with a syntax error, because the kernel Makefile newly assumes that
string values in auto.conf are not quoted.

I tested a scratch build with the cp line removed and it was possible to
build an external module without an issue. Still, to be on the safe
side, explicitly sync the timestamp of auto.conf with the root Makefile
as is done with version.h two lines above.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>