interconnect: qcom: Add SC8280XP interconnect provider
The Qualcomm SC8280XP provides the means for the OS to request bus
bandwidth using RPMh. It consists of providers for each of the busses in
the SoC and two virtual providers, for the "memory controller" and a few
of "virtual clocks".

In this initial contribution PCIe is marked keep alive, as the platform
does not yet have PCIe support and as syncstate votes for 0 bandwidth on
the PCI0 BCM the SA8540P ADP locks up. Once PCIe has been implemented
it's expected that the keepalive can be dropped again.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 32b258da59012b3ef0b31067c28817214152d74f)