Revert "PCI/MSI: Mask MSI-X vectors only on success"
This commit does indeed fix a bug, but not on systems largely used by
Fedora. It does however trigger a bug making AWS m4.large instances fail
to initialize their network.  The issue is fixed in a more modern
version of Xen, but we are reverting for now to give them time to
upgrade the m4.large hosts.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>