[redhat] New configs in drivers/nvme
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: NVME_VERBOSE_ERRORS [=n]
 Type  : bool
 Defined at drivers/nvme/host/Kconfig:27
   Prompt: NVMe verbose error reporting
   Depends on: NVME_CORE [=m]
   Location:
     Main menu
       -> Device Drivers
         -> NVME Support

---

Cc: Gopal Tiwari <gtiwari@redhat.com>
Cc: John Meneghini <jmeneghi@redhat.com>
Cc: Chris Leech <cleech@redhat.com>
Cc: "Ewan D. Milne" <emilne@redhat.com>
Cc: Maurizio Lombardi <mlombard@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>