Added a custom VFIO IOMMU implementation
The existing VFIO_IOMMU_TYPE1 driver poses an issue where the
attach_group allocates a new IOMMU domain (thus a new underlying
IOMMU context bank) for the device being used. The DMA buf framework
uses the default IOMMU domain, which was assigned during the probe of
the platform device, but VFIO uses the newly allocated DMA domain.
The SMR remains intact, but the S2CR is updated to the newly allocated CB.
All mappings from the DMA framework land on the old CB, all mappings from
the VFIO f/w land on the new CB.
The biggest issue is that the device will only see mappings in the new CB,
which means it won't see mappings from the DMA Buf f/w.

This new driver, which will be a module, circumvents that issue by not
doing any domain allocations for the VFIO f/w. It's very barebones, and
probably not the cleanest solution, but it works for the given issue.

Change-Id: I6733ff775b17ecad3eef4c592728b010f54c4ef1
Signed-off-by: Anant Goel <quic_anantg@quicinc.com>