fedora: Re-enable efifb and vesafb drivers
Unfortunately the Nvidia prorietary driver has some bugs that makes it to
not work correctly with simpledrm. Namenly the fact that doesn't register
a fbdev emulated device and so relies on existing fbdev drivers to do so.

We have a workaround that skips an "simple-framebuffer" platform device
if the nvidia-drm.modeset=1 kernel command line parameter is set and the
"efi-framebuffer" or "vesa-framebuffer" devices are registered instead.

But for this to work, the legacy efifb and vesafb drivers must be enabled.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>