[redhat] New configs in net/Kconfig.debug
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: DEBUG_NET [=n]
 Type  : bool
 Defined at net/Kconfig.debug:21
   Prompt: Add generic networking debug
   Depends on: DEBUG_KERNEL [=y]
   Location:
     Main menu
       -> Kernel hacking
         -> Networking Debugging

---

Cc: Jiri Benc <jbenc@redhat.com>
Cc: Marcelo Leitner <mleitner@redhat.com>
Cc: Andrea Claudi <aclaudi@redhat.com>
Cc: Antoine Tenart <atenart@redhat.com>
Cc: Balazs Nemeth <bnemeth@redhat.com>
Cc: Davide Caratti <dcaratti@redhat.com>
Cc: Eric Garver <egarver@redhat.com>
Cc: Flavio Leitner <fbl@redhat.com>
Cc: Florian Westphal <fwestpha@redhat.com>
Cc: Guillaume Nault <gnault@redhat.com>
Cc: Hangbin Liu <haliu@redhat.com>
Cc: Ivan Vecera <ivecera@redhat.com>
Cc: Jarod Wilson <jarod@redhat.com>
Cc: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Cc: Paolo Abeni <pabeni@redhat.com>
Cc: Phil Sutter <psutter@redhat.com>
Cc: Sabrina Dubroca <sdubroca@redhat.com>
Cc: "Toke Høiland-Jørgensen" <toke@redhat.com>
Cc: Xin Long <lxin@redhat.com>
Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>