HACK: dts: qcom: pm8150: disable RTC clock
Upstream Status: WIP only (to be removed)
Bugzilla: Internal

One of the RTC ltp tests is failing on the QDrive3 because the RTC clock
is not settable (https://issues.redhat.com/browse/VROOM-4725). It's not
clear if we are using the right RTC clock to begin with. Disable it
until we get more information on the RTC clock that is used on the
QDrive3 or until we get an upstream device tree that describes it.

This allows the RTC ltp tests to be skipped.

Signed-off-by: Adrien Thierry <athierry@redhat.com>
Signed-off-by: Andrew Halaney <ahalaney@redhat.com>