HACK: dts: qcom: pm8150: disable RTC clock
Upstream Status: WIP only (to be removed)
Bugzilla: Internal

One of the RTC ltp tests is failing on the QDrive3 because the RTC clock
is not settable (https://issues.redhat.com/browse/VROOM-4725). It's not
clear if we are using the right RTC clock to begin with. Disable it
until we get more information on the RTC clock that is used on the
QDrive3 or until we get an upstream device tree that describes it.

This allows the RTC ltp tests to be skipped.

Upstream has sc8280xp-pmic.dtsi, which we should probably be using
instead of pm8150.dtsi to begin with. Unfortunately, pm8150.dtsi
is reintroduced in ("arm64: dts: qcom: Add emac config") for
emac enablement and it's not entirely clear how to enable the same gpios
in sc8280xp-pmic.dtsi at the moment, so we're stuck with pm8150.dtsi for
the now.

Signed-off-by: Adrien Thierry <athierry@redhat.com>
Signed-off-by: Andrew Halaney <ahalaney@redhat.com>