arm64: dts: qcom: sa8540p-adp-ride: Reduce multiport to single port
Upstream Status: WIP only (to be removed)
Bugzilla: Internal

This board's multiport USB controller is only wired to one port. The SoC
level description is fine (the extra phys are unused as the dwc3 driver
looks up by name), but we shouldn't enable them and the unnecessary
gpios for this board. So remove them for now.

This patch should be squashed into
29df5bb6808d ("HACK: arm64: dts: qcom: sa8540p-adp-ride: enable USB multiport")
and the HACK should be dropped.

The real hack is
2cdfc588e775 ("arm64: dts: qcom: sc8280xp: Add multiport USB controller and phys")
which doesn't name the extra phys for multiport (since the driver
doesn't yet support using them). Not sure if upstream will accept it as
is right now, or if they'll require the extra phys to be dropped. I
think it might make sense to describe them since its the proper
description, but we'll see since dtbs_check will fail probably.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>