HACK: arm64: dts: qcom: Add sa8540p-adp-ride-lite.dts
Upstream Status: RHEL-only
Bugzilla: Internal

sa8540p-adp-ride has a bunch of configurations. One, commonly referred
to as QDrive3-lite, is a board with just one sa8540p sip (instead of
two), no sa9000p, and no PCIe switch card (just a bypass card in its
place).

As far as I know the plan longterm is only one devicetree representing
the platform devices that are hooked up since the claim is that from the
sip pov the board is the same.

For now though, PCIe dies on the full sa8540p-adp-ride (QDrive3) variant
since the PCIe switch isn't running with the right firmware. To allow
for PCIe to work on the -lite variant, add a new devicetree, which will
later enable PCIe.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>