HACK: redhat: configs: automotive: x86_64: Disable stmmac due to Qualcomm driver error
Bugzilla: INTERNAL
Upstream Status: RHEL-only

Qualcomm's latest drop for the -wip enabling stmmac for their hardware
breaks compilation with the x86 config. For the moment disable it.

Without the config change here, we get this error due to the variable
not being defined (it seems the driver drop uses compile time #ifdefs so
the stmmac driver only works with certain versions of stmmac IP -- and
won't compile for others):

    drivers/net/ethernet/stmicro/stmmac/stmmac_main.c: In function ‘stmmac_dma_reg_show’:
    drivers/net/ethernet/stmicro/stmmac/stmmac_main.c:6248:17: error: ‘V3_DMA_BUF_LEN’ undeclared (first use in this function)
     6248 |         u32 buf[V3_DMA_BUF_LEN] = {0};
          |                 ^~~~~~~~~~~~~~
    drivers/net/ethernet/stmicro/stmmac/stmmac_main.c:6248:17: note: each undeclared identifier is reported only once for each function it appears in

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>