spec: prevent git apply from searching for the .git directory
By default, git tries to find the repository root by looking for the
nearest parent directory containing a .git subdirectory. If the unpacked
tarball doesn't include the .git directory, it can lead to git trying to
apply the patch to the nearest parent git repository (e.g. the kernel
dist-git in case of a local `fedpkg prep`) instead of the current
directory.

Thus, pass `--work-tree=.` to `git apply`, so that it doesn't search for
the .git directory outside the current directory.

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>