HACK: remove mac2mac_88Q stuff, but keep it if (1)'ed
Right now, if I use fixed-link instead of mac2mac I don't get an IP
address. I assume this is because TX is borked.

With this change I do get an address, so I assume this is actually
needed in our fixed-link config. I'll need to figure out a way to
properly select that.

Unfortunately, RX side is seeing a lot of dropped packets. I need to
figure out why that is.

Save this off for now and revert back to mac2mac to see if that was
happening then too. I don't think it was.

update: I double checked, mac2mac works without the drops... so i gotta
figure that out!

update: ok now it is gone... so idk what's going on. no drops at the
moment... ill proceed cautiously.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>