hack: arm64: dts: qcom: sa8540p-ride: disable pcie2a nodes
Disable the nodes for pcie2a since we are still running the old firmware
for pcie2a. This can be dropped once we upgrade.

The previous commit adds the pcie2a node and then it's being dropped
here. The reason that I included the pcie2a change in the first place is
because the i2c/spi patches that follow this commit depend on the pcie2a
patch.

Signed-off-by: Brian Masney <bmasney@redhat.com>