HACK: Remove mac2mac in favor of fixed-link
mac2mac isn't necessary, fixed-link is the upstream equivalent.

The bit about TX phase shift however (I assume its specific to this
switch configuration) is needed for data to flow. For now keep that in
an if (1) until I find out under what scenarios it is actually
necessary, maybe we can use a dt boolean and switch off of that, or
if its a general mac2mac config need.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>