HACK: remove emac-core-version dt property and reg read
Since this is a platform device, we can encode that via the compatible.

Likewise, Qualcomm could not confirm if this register existed (I imagine
it did but have no hardware to test) on older versions. Since its not
providing any benefit let's just remove it instead of risking breaking
old hardware or encoding unnecessary info in the devicetree

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>