HACK: net: stmmac: Add EMAC3 variant of dwmac4
EMAC3 is a Qualcomm variant of dwmac4 that functions the same, but has a
different address space layout for MTL and DMA registers.

This attempt at handling that turned out ugly, I do not plan on
submitting this. I had _hoped_ that the usage was small enough that a
new variant wouldn't duplicate much code. Unfortunately, this approach
resulted in a ton of copy pasted code with only the DMA and MTL based
read/writes being different.

If you're wondering what the real change is that caused this all to be
attempted, just look at the headers. I think next I'll try and overhaul
(again, I gave up previously when trying to figure out how to deal with
all the macro functions) with regmap API calls.

I also think that since my original work and rebase at least one
function has changed, so the copy pasta now is out of date illustrating
exactly why this code isn't ready for primetime, its maintaining a bunch
of the exact same boilerplate in two places.

Signed-off-by: Andrew Halaney <ahalaney@redhat.com>